S1C2 - Movement system
Added the ability to slow down time, this will be used to great effect. It related directly to the movement and has been added to avoid overhead later. There is currently a button that slows down time and then a button that speeds it up. Slow down is bound to 'T' and normal speed is bound to 'Y'.